refactor(aipanel): 拆出搜索与会话整理 hook
- 新增 components/ai/useAIGlobalSearch.ts：全局对话搜索状态簇（开关/关键词/加载态/结果、聚焦与 180ms 防抖检索）
- 新增 components/ai/useAIConversationSearch.ts：当前对话内搜索（命中索引、结果过滤、命中定位）
- 新增 components/ai/useAIConversationOrganizer.ts：会话分组（新建/重命名/拖拽重排/删除）、多选批量操作、临时会话转正与 F2 快捷键
- AIPanel.tsx 4108 -> 3769 行
