+++ frontend/src/components/AIPanel.tsx
@@ -23,6 +23,9 @@ import { expandFirstSlashCommandForPrompt } from './ai/aiSlashCommands.ts'
 import { useAIChatStreamEvents } from './ai/useAIChatStreamEvents.ts'
 import { useAIPanelCoreState } from './ai/useAIPanelCoreState.ts'
 import { useAIPanelSettingsState } from './ai/useAIPanelSettingsState.ts'
+import { useAIGlobalSearch } from './ai/useAIGlobalSearch.ts'
+import { useAIConversationSearch } from './ai/useAIConversationSearch.ts'
+import { useAIConversationOrganizer } from './ai/useAIConversationOrganizer.ts'
 import AIChatConversation from './ai/chat/AIChatConversation.tsx'
 import { getConversationBranchAnchor } from './ai/chat/aiChatMessageTopology.ts'
 import { isCallMyVipProviderHost } from './ai/providerSpecialHosts.ts'
@@ -67,39 +70,7 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
   const [aiProviderState, setAIProviderState] = useState<AIProviderState>({ currentProviderId: '', providers: [] })
   const [isDevilMode, setIsDevilMode] = useState(false)
   const [conversationList, setConversationList] = useState<ConversationSummary[]>([])
-  const [conversationOrganizer, setConversationOrganizer] = useState<AIConversationOrganizerState>(() => loadAIConversationOrganizer())
-  const [conversationFilter, setConversationFilter] = useState('all')
-  const [conversationSelectionMode, setConversationSelectionMode] = useState(false)
-  const [selectedConversationIds, setSelectedConversationIds] = useState<Set<string>>(() => new Set())
-  const [moveToGroupOpen, setMoveToGroupOpen] = useState(false)
-  const [editingConversationGroupId, setEditingConversationGroupId] = useState('')
-  const [editingConversationGroupName, setEditingConversationGroupName] = useState('')
-  const [draggingConversationGroupId, setDraggingConversationGroupId] = useState('')
-  const [dragOverConversationGroupId, setDragOverConversationGroupId] = useState('')
-  const conversationGroupRenameInputRef = useRef<HTMLInputElement | null>(null)
-  const conversationGroupRenameCancelledRef = useRef(false)
-  const [globalSearchOpen, setGlobalSearchOpen] = useState(false)
-  const [globalSearchQuery, setGlobalSearchQuery] = useState('')
-  const [globalSearchLoading, setGlobalSearchLoading] = useState(false)
-  const [globalSearchResults, setGlobalSearchResults] = useState<AIConversationMessageSearchResult[]>([])
-  const [conversationSearchOpen, setConversationSearchOpen] = useState(false)
-  const [conversationSearchQuery, setConversationSearchQuery] = useState('')
-  const [conversationSearchIndex, setConversationSearchIndex] = useState(0)
-  const globalSearchRequestRef = useRef(0)
-  const globalSearchInputRef = useRef<HTMLInputElement | null>(null)
-  const conversationSearchInputRef = useRef<HTMLInputElement | null>(null)
-  const resetGlobalSearchState = useCallback(() => {
-    setGlobalSearchOpen(false)
-    setGlobalSearchQuery('')
-    setGlobalSearchLoading(false)
-    setGlobalSearchResults([])
-  }, [])
 
-  const resetConversationSearchState = useCallback(() => {
-    setConversationSearchOpen(false)
-    setConversationSearchQuery('')
-    setConversationSearchIndex(0)
-  }, [])
 
   const {
     panelInstanceKey, terminalPanels, setTerminalPanels, terminalPanelsRef, deletedConversationIdsRef,
@@ -113,6 +84,19 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
     getMessageApiLengthBefore, truncateConversationAfterMessage, rebuildAIConversationTokenLedger,
     refreshAIConversationContextTokens, saveConversationSnapshot,
   } = useAIPanelCoreState({ terminalId, sessionId, workspaceTabId, initialConversationId, isWorkspaceTabActive, onWorkspaceTabStateChange, setConversationList })
+
+  const {
+    globalSearchOpen, setGlobalSearchOpen, globalSearchQuery, setGlobalSearchQuery, globalSearchLoading,
+    setGlobalSearchLoading, globalSearchResults, setGlobalSearchResults, globalSearchRequestRef,
+    globalSearchInputRef, resetGlobalSearchState, normalizedGlobalSearchQuery, handleOpenGlobalSearch,
+  } = useAIGlobalSearch({ panelMountedRef })
+
+  const {
+    conversationSearchOpen, setConversationSearchOpen, conversationSearchQuery, setConversationSearchQuery,
+    conversationSearchIndex, setConversationSearchIndex, conversationSearchInputRef,
+    resetConversationSearchState, normalizedConversationSearchQuery, conversationSearchResults,
+    locateConversationMessage, handleOpenConversationSearch, handleCycleConversationSearchResult,
+  } = useAIConversationSearch({ sessionId, terminalId, workspaceTabId, panelState, activeConversation })
 
   const {
     mcpInfo, setMcpInfo, mcpClientServers, mcpClientGlobalConfigPath, mcpClientGlobalConfigText,
@@ -339,111 +323,11 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
     && !collaborationActive
     && !panelState.isCondensingContext
     && (!panelState.lastTurnBusinessMessageKind || (panelState.lastTurnBusinessMessageKind !== 'completion' && panelState.lastTurnBusinessMessageKind !== 'followup'))
-  const normalizedGlobalSearchQuery = useMemo(() => normalizeAIConversationSearchQuery(globalSearchQuery), [globalSearchQuery])
-  const normalizedConversationSearchQuery = useMemo(() => normalizeAIConversationSearchQuery(conversationSearchQuery), [conversationSearchQuery])
-  const conversationSearchResults = useMemo(() => {
-    if (!activeConversation || !normalizedConversationSearchQuery) {
-      return []
-    }
-    const normalizedNeedle = normalizedConversationSearchQuery.toLowerCase()
-    return (Array.isArray(panelState.messages) ? panelState.messages : []).flatMap((message) => {
-      const body = extractAIConversationSearchText(message)
-      if (!body || !body.toLowerCase().includes(normalizedNeedle)) {
-        return []
-      }
-      return [normalizeAIConversationMessageSearchResult({
-        conversationId: activeConversation.id,
-        conversationTitle: activeConversation.title,
-        messageId: message.id,
-        role: message.kind === 'user' ? 'user' : 'assistant',
-        snippet: buildAIConversationSearchSnippet(body, normalizedConversationSearchQuery),
-        updatedAt: activeConversation.updatedAt,
-      })]
-    })
-  }, [activeConversation, normalizedConversationSearchQuery, panelState.messages])
 
-  useEffect(() => {
-    if (!globalSearchOpen || !globalSearchInputRef.current) {
-      return
-    }
-    globalSearchInputRef.current.focus()
-    globalSearchInputRef.current.select()
-  }, [globalSearchOpen])
 
-  useEffect(() => {
-    if (!conversationSearchOpen || !conversationSearchInputRef.current) {
-      return
-    }
-    conversationSearchInputRef.current.focus()
-    conversationSearchInputRef.current.select()
-  }, [conversationSearchOpen])
 
-  useEffect(() => {
-    if (!conversationSearchOpen) {
-      return
-    }
-    if (conversationSearchResults.length === 0) {
-      setConversationSearchIndex(0)
-      return
-    }
-    setConversationSearchIndex((current) => (current >= conversationSearchResults.length ? 0 : current))
-  }, [conversationSearchOpen, conversationSearchResults.length])
 
-  useEffect(() => {
-    if (!conversationSearchOpen || !normalizedConversationSearchQuery || conversationSearchResults.length === 0) {
-      return
-    }
-    const activeResult = conversationSearchResults[conversationSearchIndex] || conversationSearchResults[0]
-    if (!activeResult?.messageId || typeof window === 'undefined') {
-      return
-    }
-    window.dispatchEvent(new CustomEvent('ai-conversation-diff-locate', {
-      detail: {
-        sessionId: sessionId || '',
-        terminalId: terminalId || '',
-        tabId: workspaceTabId,
-        messageId: activeResult.messageId,
-      },
-    }))
-  }, [conversationSearchIndex, conversationSearchOpen, conversationSearchResults, normalizedConversationSearchQuery, sessionId, terminalId])
 
-  useEffect(() => {
-    if (!globalSearchOpen) {
-      setGlobalSearchLoading(false)
-      setGlobalSearchResults([])
-      return
-    }
-    if (!normalizedGlobalSearchQuery) {
-      setGlobalSearchLoading(false)
-      setGlobalSearchResults([])
-      return
-    }
-    const requestId = globalSearchRequestRef.current + 1
-    globalSearchRequestRef.current = requestId
-    setGlobalSearchLoading(true)
-    const timer = window.setTimeout(() => {
-      searchAIConversationMessages(normalizedGlobalSearchQuery, '', 50)
-        .then((results) => {
-          if (!panelMountedRef.current || globalSearchRequestRef.current !== requestId) {
-            return
-          }
-          setGlobalSearchResults(results)
-        })
-        .catch(() => {
-          if (!panelMountedRef.current || globalSearchRequestRef.current !== requestId) {
-            return
-          }
-          setGlobalSearchResults([])
-        })
-        .finally(() => {
-          if (!panelMountedRef.current || globalSearchRequestRef.current !== requestId) {
-            return
-          }
-          setGlobalSearchLoading(false)
-        })
-    }, 180)
-    return () => window.clearTimeout(timer)
-  }, [globalSearchOpen, normalizedGlobalSearchQuery])
 
 
 
@@ -853,53 +737,8 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
     addToast?.(t('任务标题已更新'), 'success')
   }, [activeConversation, addToast, panelInstanceKey, saveConversationSnapshot, setPanelState, t])
 
-  const locateConversationMessage = useCallback((messageId: string) => {
-    const normalizedMessageId = typeof messageId === 'string' ? messageId.trim() : ''
-    if (!normalizedMessageId || typeof window === 'undefined') {
-      return
-    }
-    window.dispatchEvent(new CustomEvent('ai-conversation-diff-locate', {
-      detail: {
-        sessionId: sessionId || '',
-        terminalId: terminalId || '',
-        tabId: workspaceTabId,
-        messageId: normalizedMessageId,
-      },
-    }))
-  }, [sessionId, terminalId, workspaceTabId])
 
-  const handleOpenGlobalSearch = useCallback(() => {
-    setGlobalSearchOpen((current) => {
-      const next = !current
-      if (!next) {
-        setGlobalSearchQuery('')
-        setGlobalSearchLoading(false)
-        setGlobalSearchResults([])
-      }
-      return next
-    })
-  }, [])
 
-  const handleOpenConversationSearch = useCallback(() => {
-    setConversationSearchOpen((current) => {
-      const next = !current
-      if (!next) {
-        setConversationSearchQuery('')
-        setConversationSearchIndex(0)
-      }
-      return next
-    })
-  }, [])
-
-  const handleCycleConversationSearchResult = useCallback((direction: number) => {
-    if (conversationSearchResults.length === 0) {
-      return
-    }
-    setConversationSearchIndex((current) => {
-      const total = conversationSearchResults.length
-      return (current + direction + total) % total
-    })
-  }, [conversationSearchResults.length])
 
   const handleSelectGlobalSearchResult = useCallback(async (result: AIConversationMessageSearchResult) => {
     const conversationId = typeof result?.conversationId === 'string' ? result.conversationId.trim() : ''
@@ -956,157 +795,17 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
     }
   }, [clearRestorePreview, handleGoHome, handleOpenConversation, panelInstanceKey, panelState.activeConversationId, requestDeleteConfirmation, t])
 
-  const persistConversationOrganizer = useCallback((updater: (current: AIConversationOrganizerState) => AIConversationOrganizerState) => {
-    setConversationOrganizer((current) => saveAIConversationOrganizer(updater(current)))
-  }, [])
 
-  const handleMakeConversationPermanent = useCallback(async (conversationId: string) => {
-    const temporarySummary = getTemporaryAIConversationSummary(conversationId)
-    if (!temporarySummary) return
-    let createdConversationId = ''
-    try {
-      const temporarySnapshot = await getTemporaryAIConversation(conversationId)
-      const created = await createAIConversation(temporarySnapshot.title || t('新对话'))
-      createdConversationId = created.id
-      const permanentSnapshot = await saveAIConversation({
-        ...temporarySnapshot,
-        id: created.id,
-        createdAt: created.createdAt,
-        updatedAt: Date.now(),
-        transient: false,
-        status: 'idle',
-      })
-      removeTemporaryAIConversation(conversationId)
-      persistConversationOrganizer((current) => {
-        const assignments = { ...current.assignments }
-        if (assignments[conversationId]) assignments[permanentSnapshot.id] = assignments[conversationId]
-        delete assignments[conversationId]
-        return { ...current, assignments }
-      })
-      setConversationList((current) => upsertConversationSummary(current.filter((item) => item.id !== conversationId), permanentSnapshot))
-      addToast?.(`${t('临时会话')} · ${t('保存')}`, 'success')
-      await handleOpenConversation(permanentSnapshot.id)
-    } catch {
-      if (createdConversationId) await deleteAIConversation(createdConversationId).catch(() => {})
-      addToast?.(t('保存失败'), 'error')
-    }
-  }, [addToast, handleOpenConversation, persistConversationOrganizer, t])
 
-  const handleCreateConversationGroup = useCallback(async () => {
-    const name = window?.luminDialog?.prompt
-      ? await window.luminDialog.prompt(t('请输入分组名称'), '', t('新建分组'))
-      : window.prompt(t('请输入分组名称'))
-    const normalizedName = typeof name === 'string' ? name.trim() : ''
-    if (!normalizedName) return
-    const group = createAIConversationGroup(normalizedName)
-    persistConversationOrganizer((current) => ({ ...current, groups: [...current.groups, group] }))
-    setConversationFilter(group.id)
-  }, [persistConversationOrganizer, t])
 
-  const beginRenameConversationGroup = useCallback((groupId: string) => {
-    const group = conversationOrganizer.groups.find((item) => item.id === groupId)
-    if (!group) return
-    setConversationFilter(groupId)
-    conversationGroupRenameCancelledRef.current = false
-    setEditingConversationGroupId(groupId)
-    setEditingConversationGroupName(group.name)
-    window.requestAnimationFrame(() => {
-      conversationGroupRenameInputRef.current?.focus()
-      conversationGroupRenameInputRef.current?.select()
-    })
-  }, [conversationOrganizer.groups])
 
-  const cancelRenameConversationGroup = useCallback(() => {
-    conversationGroupRenameCancelledRef.current = true
-    setEditingConversationGroupId('')
-    setEditingConversationGroupName('')
-  }, [])
 
-  const commitRenameConversationGroup = useCallback(() => {
-    if (conversationGroupRenameCancelledRef.current) {
-      conversationGroupRenameCancelledRef.current = false
-      setEditingConversationGroupId('')
-      setEditingConversationGroupName('')
-      return
-    }
-    const groupId = editingConversationGroupId
-    const normalizedName = editingConversationGroupName.trim()
-    if (groupId && normalizedName) {
-      persistConversationOrganizer((current) => ({ ...current, groups: current.groups.map((item) => item.id === groupId ? { ...item, name: normalizedName } : item) }))
-    }
-    setEditingConversationGroupId('')
-    setEditingConversationGroupName('')
-  }, [editingConversationGroupId, editingConversationGroupName, persistConversationOrganizer])
 
-  const reorderConversationGroup = useCallback((sourceGroupId: string, targetGroupId: string) => {
-    if (!sourceGroupId || !targetGroupId || sourceGroupId === targetGroupId) return
-    persistConversationOrganizer((current) => {
-      const sourceIndex = current.groups.findIndex((group) => group.id === sourceGroupId)
-      const targetIndex = current.groups.findIndex((group) => group.id === targetGroupId)
-      if (sourceIndex < 0 || targetIndex < 0) return current
-      const groups = [...current.groups]
-      const [moved] = groups.splice(sourceIndex, 1)
-      groups.splice(targetIndex, 0, moved)
-      return { ...current, groups }
-    })
-  }, [persistConversationOrganizer])
 
-  const showSystemGroupRenameUnsupported = useCallback(() => {
-    const message = t('系统分组不支持重命名')
-    if (typeof addToast === 'function') {
-      addToast(message, 'info', 2200)
-      return
-    }
-    void showAlert(message)
-  }, [addToast, showAlert, t])
 
-  useEffect(() => {
-    if (!isWorkspaceTabActive) return
-    const handleRenameShortcut = (event: KeyboardEvent) => {
-      if (event.key !== 'F2' || editingConversationGroupId) return
-      if (conversationFilter === 'all' || conversationFilter === 'archived') {
-        event.preventDefault()
-        showSystemGroupRenameUnsupported()
-        return
-      }
-      if (!conversationOrganizer.groups.some((group) => group.id === conversationFilter)) return
-      event.preventDefault()
-      beginRenameConversationGroup(conversationFilter)
-    }
-    window.addEventListener('keydown', handleRenameShortcut)
-    return () => window.removeEventListener('keydown', handleRenameShortcut)
-  }, [beginRenameConversationGroup, conversationFilter, conversationOrganizer.groups, editingConversationGroupId, isWorkspaceTabActive, showSystemGroupRenameUnsupported])
 
-  useEffect(() => {
-    if (conversationFilter === 'ungrouped') setConversationFilter('all')
-  }, [conversationFilter])
 
-  const handleDeleteConversationGroup = useCallback(async (groupId: string) => {
-    const group = conversationOrganizer.groups.find((item) => item.id === groupId)
-    if (!group) return
-    const confirmed = await requestDeleteConfirmation(t('删除分组后,其中的会话将移到未分组.是否继续?'))
-    if (!confirmed) return
-    persistConversationOrganizer((current) => ({
-      groups: current.groups.filter((item) => item.id !== groupId),
-      assignments: Object.fromEntries(Object.entries(current.assignments).filter(([, assignedGroupId]) => assignedGroupId !== groupId)),
-    }))
-    if (conversationFilter === groupId) setConversationFilter('ungrouped')
-  }, [conversationFilter, conversationOrganizer.groups, persistConversationOrganizer, requestDeleteConfirmation, t])
 
-  const toggleConversationSelection = useCallback((conversationId: string) => {
-    setSelectedConversationIds((current) => {
-      const next = new Set(current)
-      if (next.has(conversationId)) next.delete(conversationId)
-      else next.add(conversationId)
-      return next
-    })
-  }, [])
-
-  const clearConversationSelection = useCallback(() => {
-    setSelectedConversationIds(new Set())
-    setConversationSelectionMode(false)
-    setMoveToGroupOpen(false)
-  }, [])
 
   const refreshConversationList = useCallback(async () => {
     const conversations = await listAIConversations().catch(() => [])
@@ -1114,59 +813,21 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
     seedTemporaryAIConversations(temporarySummaries)
     setConversationList([...temporarySummaries.map((summary) => ({ ...summary, transient: true })), ...(Array.isArray(conversations) ? conversations : [])])
   }, [])
+  const {
+    conversationOrganizer, setConversationOrganizer, conversationFilter, setConversationFilter,
+    conversationSelectionMode, setConversationSelectionMode, selectedConversationIds, setSelectedConversationIds,
+    moveToGroupOpen, setMoveToGroupOpen, editingConversationGroupId, setEditingConversationGroupId,
+    editingConversationGroupName, setEditingConversationGroupName, draggingConversationGroupId,
+    setDraggingConversationGroupId, dragOverConversationGroupId, setDragOverConversationGroupId,
+    conversationGroupRenameInputRef, conversationGroupRenameCancelledRef, persistConversationOrganizer,
+    handleMakeConversationPermanent, handleCreateConversationGroup, beginRenameConversationGroup,
+    cancelRenameConversationGroup, commitRenameConversationGroup, reorderConversationGroup,
+    showSystemGroupRenameUnsupported, handleDeleteConversationGroup, toggleConversationSelection,
+    clearConversationSelection, handleMoveSelectedConversations, handleSetSelectedArchived,
+    handleDeleteSelectedConversations,
+  } = useAIConversationOrganizer({ t, addToast, showAlert, requestDeleteConfirmation, isWorkspaceTabActive, refreshConversationList, handleOpenConversation, setConversationList })
 
-  const handleMoveSelectedConversations = useCallback((groupId: string) => {
-    const selected = new Set(selectedConversationIds)
-    persistConversationOrganizer((current) => {
-      const assignments = { ...current.assignments }
-      selected.forEach((conversationId) => {
-        if (groupId) assignments[conversationId] = groupId
-        else delete assignments[conversationId]
-      })
-      return { ...current, assignments }
-    })
-    clearConversationSelection()
-  }, [clearConversationSelection, persistConversationOrganizer, selectedConversationIds])
 
-  const handleSetSelectedArchived = useCallback(async (archived: boolean) => {
-    const ids = Array.from(selectedConversationIds)
-    await Promise.all(ids.map(async (conversationId) => {
-      try {
-    const temporarySummary = getTemporaryAIConversationSummary(conversationId)
-        if (temporarySummary) {
-          const temporarySnapshot = await getTemporaryAIConversation(conversationId)
-          const saved = await saveTemporaryAIConversation({ ...temporarySnapshot, archived, updatedAt: Date.now() })
-          upsertTemporaryAIConversation(saved)
-          return
-        }
-        const snapshot = await getAIConversation(conversationId)
-        await saveAIConversation({ ...snapshot, archived, updatedAt: Date.now() })
-      } catch {
-        // Continue processing the remaining selected conversations.
-      }
-    }))
-    clearConversationSelection()
-    await refreshConversationList()
-  }, [clearConversationSelection, refreshConversationList, selectedConversationIds])
-
-  const handleDeleteSelectedConversations = useCallback(async () => {
-    const ids = Array.from(selectedConversationIds)
-    if (ids.length === 0) return
-    const confirmed = await requestDeleteConfirmation(t('确定删除选中的对话吗？此操作不可撤销。'))
-    if (!confirmed) return
-    const results = await Promise.allSettled(ids.map(async (conversationId) => removeTemporaryAIConversation(conversationId) ? deleteTemporaryAIConversation(conversationId) : deleteAIConversation(conversationId)))
-    persistConversationOrganizer((current) => ({
-      ...current,
-      assignments: Object.fromEntries(Object.entries(current.assignments).filter(([conversationId]) => !selectedConversationIds.has(conversationId))),
-    }))
-    clearConversationSelection()
-    await refreshConversationList()
-    // 容错：部分失败时提示，不影响已成功的删除
-    const failedCount = results.filter((result) => result.status === 'rejected').length
-    if (failedCount > 0) {
-      addToast?.(`${t('部分对话删除失败')}（${failedCount}），其余删除已生效`, 'error')
-    }
-  }, [addToast, clearConversationSelection, persistConversationOrganizer, refreshConversationList, requestDeleteConfirmation, selectedConversationIds, t])
   useEffect(() => {
     const handleDeleteWorkspaceTabConversation = (event: Event) => {
       const detail = (event as CustomEvent).detail || {}
